Resolve file for `magic_word`.
If the magic word file is relative and not in the submission root, and
the submission root doesn't exist, we might get a weird situation where
the file exists but the grader claims it doesn't. In this case, we
instead resolve the file path to hop out of the root.